Rename modules and cleanup.
